Prune old implementForallIntents1,2
